[design/#142] extension 파일 추가
